Add files via upload
button_authentication.py is a GUI to test retinal authentication and password input.

It functions as the following:

When the mouse hovers over one of the buttons 1-9, it takes that as input each time entering it in as a sequence. Once you have created your password sequence you can than hit 'Authenticate' to test your sequence against the password sequence. If the inputted sequence is the password sequence than your authentication attempt is a success, otherwise a failure.